part4 through exercise 4.21

+++ controllers/blogs.js
@@ -1,19 +1,20 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
-const logger = require('../utils/logger')
+
+const jwt = require('jsonwebtoken')
+
+const User = require('../models/user')
 
 blogsRouter.get('/', async (request, response, next) => {
 
   try {
-    const blogs = await Blog.find({})
+    const blogs = await Blog
+      .find({}).populate('user', { username: 1, name: 1 })
+
     response.json(blogs.map(blog => blog.toJSON()))
   } catch (exception) {
     next(exception)
   }
-
-  // Blog.find({}).then(blogs => {
-  //   response.json(blogs.map(blog => blog.toJSON()))
-  // })
 })
 
 blogsRouter.get('/:id', async (request, response, next) => {
@@ -32,27 +33,57 @@ blogsRouter.get('/:id', async (request, response, next) => {
 
 blogsRouter.post('/', async (request, response, next) => {
 
-  const blog = new Blog(request.body)
+  const body = request.body
 
   try {
-    const savedNote = await blog.save()
-    response.json(savedNote.toJSON())
+
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!request.token || !decodedToken.id) {
+      return response.status(401).json({ error: 'token missing or invalid' })
+    }
+
+    const user = await User.findById(decodedToken.id)
+
+    const blog = new Blog({
+      title: body.title,
+      author: body.author === undefined ? null : body.author,
+      url: body.url,
+      likes: body.likes,
+      user: user._id
+    })
+
+
+    const savedBlog = await blog.save()
+    user.blogs = user.blogs.concat(savedBlog._id)
+    await user.save()
+    response.json(savedBlog.toJSON())
   } catch (exception) {
     next(exception)
   }
-
-  // blog.save()
-  //   .then(savedBlog => {
-  //     response.json(savedBlog.toJSON())
-  //   })
-  //   .catch(error => next(error))
 })
 
 blogsRouter.delete('/:id', async (request, response, next) => {
 
   try {
-    await Blog.findByIdAndRemove(request.params.id)
-    response.status(204).end()
+    const blog = await Blog.findById(request.params.id)
+
+    if (!blog){
+      response.send(400).end()
+    }
+
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!request.token || !decodedToken.id) {
+      return response.status(401).json({ error: 'token missing or invalid' })
+    }
+
+    const user = await User.findById(decodedToken.id)
+
+    if (blog.user.toString() === user.id.toString() ) {
+      await Blog.findByIdAndRemove(request.params.id)
+      response.status(204).end()
+    } else {
+      response.status(401).json({ error: 'you are not the author of this blog' })
+    }
   } catch (exception) {
     next(exception)
   }
@@ -75,12 +106,6 @@ blogsRouter.put('/:id', async (request, response, next) => {
   } catch(exception) {
     next(exception)
   }
-
-  // Blog.findByIdAndUpdate(request.params.id, blog, { new: true })
-  //   .then(updatedBlog => {
-  //     response.json(updatedBlog.toJSON())
-  //   })
-  //   .catch(error => next(error))
 })
 
 module.exports = blogsRouter
